+++ starsky/starsky/Startup.cs
@@ -103,18 +103,15 @@ namespace starsky
             {
                 case (AppSettings.DatabaseTypeList.Mysql):
                     services.AddDbContext<ApplicationDbContext>(options => options.UseMySql(_appSettings.DatabaseConnection, 
-	                    b => b.MigrationsAssembly(foundationDatabaseName)),
-	                    ServiceLifetime.Transient);
+	                    b => b.MigrationsAssembly(foundationDatabaseName)));
                     services.AddHealthChecks().AddMySql(_appSettings.DatabaseConnection);
                     break;
                 case AppSettings.DatabaseTypeList.InMemoryDatabase:
-                    services.AddDbContext<ApplicationDbContext>(options => options.UseInMemoryDatabase("starsky"),
-	                    ServiceLifetime.Transient);
+                    services.AddDbContext<ApplicationDbContext>(options => options.UseInMemoryDatabase("starsky"));
                     break;
                 case AppSettings.DatabaseTypeList.Sqlite:
                     services.AddDbContext<ApplicationDbContext>(options => options.UseSqlite(_appSettings.DatabaseConnection, 
-	                    b => b.MigrationsAssembly(foundationDatabaseName)),
-	                    ServiceLifetime.Transient);
+	                    b => b.MigrationsAssembly(foundationDatabaseName)));
                     services.AddHealthChecks().AddSqlite(_appSettings.DatabaseConnection, healthSqlQuery, "sqlite");
                     break;
             }