+++ starsky/starskytest/starsky.foundation.storage/Services/StructureServiceTest.cs
@@ -19,7 +19,7 @@ namespace starskytest.starsky.foundation.storage.Services
 		{
 			var structure = "/yyyy/MM/yyyy_MM_dd/yyyyMMdd_HHmmss.ext";
 			var importItem = new StructureService(new FakeIStorage(), structure);
-			var fileName = importItem.ParseFileName(new DateTime(), string.Empty, ExtensionRolesHelper.ImageFormat.jpg );
+			var fileName = importItem.ParseFileName(new DateTime(), string.Empty, "jpg" );
 			Assert.AreEqual("00010101_000000.jpg", fileName);
 		}
 		
@@ -28,7 +28,7 @@ namespace starskytest.starsky.foundation.storage.Services
 		{
 			var structure = "/yyyyMMdd_HHmmss_\\\\\\h\\\\\\m.ext";
 			var importItem = new StructureService(new FakeIStorage(), structure);
-			var fileName = importItem.ParseFileName(new DateTime(), string.Empty, ExtensionRolesHelper.ImageFormat.jpg );
+			var fileName = importItem.ParseFileName(new DateTime(), string.Empty, "jpg" );
 			Assert.AreEqual("00010101_000000_hm.jpg", fileName);
 		}
 		
@@ -37,7 +37,7 @@ namespace starskytest.starsky.foundation.storage.Services
 		{
 			var structure = "/yyyy/MM/yyyy_MM_dd/yyyyMMdd_HHmmss_\\d.ext";
 			var importItem = new StructureService(new FakeIStorage(), structure);
-			var fileName = importItem.ParseFileName(new DateTime(), string.Empty, ExtensionRolesHelper.ImageFormat.jpg );
+			var fileName = importItem.ParseFileName(new DateTime(), string.Empty, "jpg" );
 			Assert.AreEqual("00010101_000000_d.jpg", fileName);
 		}
 
@@ -48,7 +48,7 @@ namespace starskytest.starsky.foundation.storage.Services
 			var importItem = new StructureService(new FakeIStorage(), structure);
 			var fileName = importItem.ParseFileName(
 				new DateTime(2020, 01, 01, 01, 01, 01), 
-				"test", ExtensionRolesHelper.ImageFormat.jpg );
+				"test", "jpg" );
 			
 			Assert.AreEqual("20200101_010101_test.jpg", fileName);
 		}